% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vline_01.R
\name{geom_vline_01}
\alias{geom_vline_01}
\title{Ggplot vertical marker lines with default settings}
\usage{
geom_vline_01(...)
}
\description{
Ggplot vertical marker lines with default settings: alpha = 0.2, linetype = "dashed"
}
\details{
Description
}
\examples{
p <- ggplot(df_example, aes(id, value)) + geom_point()
p + geom_vline(c(5))
}
