% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_decode.R
\name{html_decode}
\alias{html_decode}
\title{Decode a character vector}
\usage{
html_decode(str)
}
\arguments{
\item{str}{A character vector}
}
\value{
Decoded character vector
}
\description{
This function is a wrapper around existing C++ code on decoding HTML entities.
The original C++ code is given by Christoph. Please refer to the answer on SO
\href{https://stackoverflow.com/a/1082191/10437891}{here}.
}
\examples{
html_decode(c("&amp;", "&euro;"))
}
