% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xp_01.R
\name{theme_xp_01}
\alias{theme_xp_01}
\title{Ggplot template 01}
\usage{
theme_xp_01(rotate = FALSE, legend.position = "bottom")
}
\description{
Personalised ggplot template 01
}
\details{
Description
}
\examples{
p <- ggplot(df_example, aes(id, value)) + geom_point()
p + theme_xp_01()
p + theme_xp_01(rotate = TRUE)
}
